+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 1.1.1 - 2022-04-14 (Pre-release for 1.2.0)
+
+### Breaking Changes:
+
+- Configuration now required to use Abaco with docker from the get go.
+
+### New features:
+
+- New Makefile with ability to manage Kubernetes
+- New code organization
+- Abaco can now use Kubernetes as a backend. This requires deployment on minikube and setting config.json `backend_container=kubernetes`.
+
+### Bug fixes:
+
+- No Change.
+
 
 ## 1.0.2 - 2022-01-06 (Pre-release for 1.1.0)
 
@@ -15,6 +31,7 @@ All notable changes to this project will be documented in this file.
 ### Bug fixes:
 
 - No Change.
+
 
 ## 1.0.0 - 2021-09-24